move ecs files into ecs directory
