+++ src/ecs/systems/ai.system.js
@@ -1,6 +1,6 @@
 import { addComponent } from "bitecs";
-import { MoveTo, Position } from "../ecs/components";
-import { aStar } from "../lib/pathfinding";
+import { MoveTo, Position } from "../components";
+import { aStar } from "../../lib/pathfinding";
 import { aiQuery, pcQuery } from "../queries";
 
 const moveToTarget = (world, eid, targetEid) => {
