+++ src/ecs/systems/movement.system.js
@@ -1,6 +1,6 @@
 import _ from "lodash";
 import { addComponent, hasComponent, removeComponent } from "bitecs";
-import { addLog } from "../index";
+import { addLog } from "../../index";
 import {
   Ai,
   Blocking,
@@ -13,10 +13,10 @@ import {
   Strength,
   MoveTo,
   Zindex,
-} from "../ecs/components";
-import { getWielders, updatePosition } from "../ecs/ecsHelpers";
-import { grid } from "../lib/grid";
-import * as gfx from "../lib/graphics";
+} from "../components";
+import { getWielders, updatePosition } from "../ecsHelpers";
+import { grid } from "../../lib/grid";
+import * as gfx from "../../lib/graphics";
 import { movementQuery, pcQuery } from "../queries";
 
 const kill = (world, eid) => {
